Don't crash when using pytester in default mode
If you try to test a pytest plugin using the pytester module, and you
also have pytest-faulthandler installed, then currently it crashes
with obscure errors. The problem is that by default, pytester runs the
sub-pytest in-process, and the sub-pytest has sys.stderr set to a
_pytest.capture.CaptureIO object, which has a fileno method (so the
old hasattr(sys.stderr, "fileno") check passes), but if you try to
call this method then it raises io.UnsupportedOperation.

So this commit robustifies the sys.stderr.fileno() calling code to
handle both a missing fileno (as before), and it raising
UnsupportedOperation.